% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_rows.R
\name{dup_rows}
\alias{dup_rows}
\title{Find all duplicate rows in provided columns and sort the rows by the columns.}
\usage{
dup_rows(data, ...)
}
\description{
Note that this function provides all duplicate rows including the first one in each duplicate group.
}
\examples{
library(tidyverse)
iris \%>\% dup_rows(Sepal.Length, Sepal.Width)
}
\keyword{duplicates}
