% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_by.R
\name{duplicates_by}
\alias{duplicates_by}
\title{Find all duplicate rows by chosen columns.}
\usage{
duplicates_by(data, ..., sort = TRUE)
}
\arguments{
\item{data}{a tbl}

\item{...}{Variables to check duplicates by.}

\item{sort}{When sort=TRUE as in the default, the output will be sorted by the variables provided in ...}
}
\description{
Note that this function generates all duplicate rows including the first one in each duplicate group.
}
\examples{
library(tidyverse)
iris \%>\% duplicates_by(Sepal.Length, Sepal.Width)
}
\keyword{duplicates}
