% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_m.R
\name{left_join_m}
\alias{left_join_m}
\title{left_join with messages}
\usage{
left_join_m(df_left, df_right, by = NULL, ...)
}
\description{
This function runs left_join and provides information on match type (1:1, 1:m, m:1, m:m) and the number of unmatched rows in the left tibble.
}
\examples{
library(tidyverse)
band_members \%>\% left_join_m(band_instruments)
band_members \%>\% left_join_m(band_instruments, by = "name")
band_members \%>\% left_join_m(band_instruments2, by = c("name" = "artist"))
}
\keyword{left_join}
