% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cluster.R
\name{lm_cluster}
\alias{lm_cluster}
\title{Fitting Linear Models with Clustered Errors}
\usage{
lm_cluster(data, formula, cluster, ...)
}
\description{
This function calculates standard errors with clustered erros and add them to the linear model fit object with a named vector "se_cluster".
This vector can be used in stargazer.
}
\examples{

## Petersen's data
data("PetersenCL", package = "sandwich")
m <- lm(y ~ x, data = PetersenCL)

## clustered covariances
m1 <- lm_cluster(y ~ x, data = PetersenCL, cluster=firm)
summary(m1)
m1$se_cluster
m1$vcov_cluster

stargazer(m, m1, se=list(NULL, m1$se_cluster))
}
\keyword{clustered}
\keyword{errors}
\keyword{standard}
\keyword{with}
