% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_mean.R
\name{se_mean}
\alias{se_mean}
\title{Standard error of mean, with na.rm option}
\usage{
se_mean(x, na.rm = FALSE)
}
\description{
This function calculates the standard error of mean, with na.rm option
}
\examples{
mtcars \%>\% group_by(cyl, vs) \%>\% summarise_combo(cyl_n = n(), mean(mpg), se_mean(mpg))
}
\keyword{error,}
\keyword{mean}
\keyword{standard}
