% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise2.R
\name{summarise2}
\alias{summarise2}
\title{summarise for each combination of group_by variables}
\usage{
summarise2(data, ...)
}
\description{
This function extends dplyr::summarise to generate summaries for each combination of group_by variables
}
\examples{
mtcars \%>\%
group_by(cyl, vs) \%>\%
summarise2(cyl_n = n())
}
\keyword{summarise}
