% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_combo.R
\name{summarise_combo}
\alias{summarise_combo}
\title{Run 'dplyr::summarise' for each combination of group_by variables}
\usage{
summarise_combo(data, ..., type = FALSE, na.str = TRUE)
}
\arguments{
\item{data}{data frame or tibble}

\item{...}{summarise commands}

\item{type}{If TRUE, it adds "_type_" column, an identifier for each group.}

\item{na.str}{If TRUE, it convers NA to a string "NA" within each group. This distinguished "NA" as a group variable value from NA indicating the variable was not used as a group variable. See Examples.}
}
\description{
This function runs dplyr::summarise for each combination of group variables.
}
\details{
https://stackoverflow.com/questions/28992028/grouping-over-all-possible-combinations-of-several-variables-with-dplyr/47949800#47949800
}
\examples{
library(tidyverse)
data <- tibble(a = c("a1","a1","a2","a3"), b= c("b1", "b2", "b2", NA), c=c(1,2,3,4) )
data \%>\% group_by(a,b) \%>\% summarise_combo(n(), mean(c))
data \%>\% group_by(a,b) \%>\% summarise_combo(n(), mean(c), type=TRUE)
data \%>\% group_by(a,b) \%>\% summarise_combo(n(), mean(c), type=TRUE, na.str=FALSE)
}
\keyword{combo}
\keyword{summarise,}
