% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_null2na.R
\name{unlist_null2na}
\alias{unlist_null2na}
\title{unlist_null2na}
\usage{
unlist_null2na(l, ...)
}
\arguments{
\item{...}{passed to unlist()}

\item{x}{input list}
}
\description{
This function converts NULL to NA and then apply unlist().
Inspired by https://stackoverflow.com/questions/2991514/prevent-unlist-to-drop-null-values
}
\examples{
x <- list(NULL, 1, 2)
unlist_null2na(x)
}
\keyword{NA}
\keyword{NULL,}
\keyword{unlist,}
