% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{classify}
\usage{
classify(x, n, style = "equal", num_format = scales::comma_format())
}
\description{
This function discretize a numeric vector into an interval factor
}
\examples{
library(ggplot2)
library(sf)
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
ggplot(nc) +  geom_sf(aes(fill = classify(AREA, 3, style ="jenks", num_format = scales::comma_format(accuracy = 0.01))))
}
\keyword{geom_sf}
