% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{classify}
\usage{
classify(x, n, style = "equal", label = scales::comma_format())
}
\arguments{
\item{x}{input vector (numeric)}

\item{n}{number of intervals}

\item{style}{"equal", "quantile", "jenks"}

\item{label}{a vectorized function that formats numbers.}
}
\description{
This function discretize a numeric vector into an interval factor
}
\examples{
library(ggplot2)
library(sf)
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
ggplot(nc) +  geom_sf(aes(fill = classify(AREA, 3, style ="jenks", label = scales::comma_format(accuracy = 0.01))))
}
\keyword{Jenks,}
\keyword{breaks,}
\keyword{classification}
\keyword{equal}
\keyword{geom_sf,}
\keyword{interval,}
\keyword{natural}
\keyword{quantile,}
