% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_combo.R
\name{summarise_combo}
\alias{summarise_combo}
\title{Run 'dplyr::summarise' for each combination of group_by variables}
\usage{
summarise_combo(data, ...)
}
\description{
This function runs dplyr::summarise for each combination of group variables.
}
\details{
https://stackoverflow.com/questions/28992028/grouping-over-all-possible-combinations-of-several-variables-with-dplyr/47949800#47949800
}
\examples{
library(tidyverse)
mtcars \%>\% group_by(cyl, vs) \%>\% summarise_combo(cyl_n = n(), mean(mpg))
}
\keyword{summarise}
